refactor(hooks): edit export statements
