+++ src/hooks/useIntersectionObserver/index.ts
@@ -1,1 +1,1 @@
-export * from './useIntersectionObserver';
+export { useIntersectionObserver } from './useIntersectionObserver';
